requirements/dev.txt
-r common.txt

pytest
pytest-django
pytest-testdox
pytest-asyncio

django-debug-toolbar